wip(baseline): move types around to check:tsc works
